backend/requirements.txt
djangorestframework>=3.14
django-cors-headers>=4.3
